Clean up rules
I used `systat rules` to see which rules were getting hit. I also
noticed that putting `{}` inline in rules makes the rule expand to
multiple rules, so I switched some of those to tables. The ext_ifs rules
are now split out, so that it's easier to apply queuing to them.